[ODM-11711] Move intializer application to filesUtil
